tiles, skybox, some improvements to things, mostly wip stuff in this commit

+++ source/ogg_vorbis.hpp
@@ -26,7 +26,7 @@ private:
 	io_stream file_stream;
 	io_stream pcm_stream;
 
-	long frequency{ 0 };
+	int frequency{ 0 };
 	int channels{ 0 };
 
 };
